+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Add stream signer.
+
 ## 0.1.0 (2023-10-25)
 
 - Initial release.